// src/types/Harassment.ts
export type Harassment = {
  age: number;
  salary: number;
  jobTitle: string;
  department: string;
  years: number;
  amount: string;

  health:string;
  unfit:string;


  workDepartment:string;
  employees:number;
  reason:string;

  committed:string;
  splRelationship:string;

  moralElements:[string];

  evidenceTypes:[string];
  alleged :string;
  otherEvidence:string;
  otherSpecify: string ,

  damages:[string];

  situation:string;

  progress: number;
  nextProgress: number;
  checked: boolean;

  user: string;
  caseId: string;
};